Clean up decode case and guards
It shouldn't be possible to pass a malformed binary to decode and get
a badarg exception anymore now that it's checked that Bin is at least
as large as the size of the message it contains (if it contains a valid
message).

decode_with_type has been moved down a bit and and
decode_return_with_rest accounts for returning the remainder of the
initial binary.